requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
coverage==5.5
Django==3.1
djangorestframework==3.12.2
Faker==6.6.0
importlib-metadata==3.7.2
iniconfig==1.1.1
mixer==7.1.2
numpy==1.20.1
packaging==20.9
pandas==1.2.3
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
zipp==3.4.1
